+++ packages/components/button/src/variants/ghost.ts
@@ -6,9 +6,9 @@ export const ghostVariants = [
     design: 'ghost',
     class: tw([
       'text-primary',
-      'enabled:hover:bg-primary-container-hovered',
-      'enabled:active:bg-primary-container-hovered',
-      'focus-visible:bg-primary-container-focused',
+      'enabled:hover:bg-primary/dim-5',
+      'enabled:active:bg-primary/dim-5',
+      'focus-visible:bg-primary/dim-5',
     ]),
   },
   {
@@ -16,9 +16,9 @@ export const ghostVariants = [
     design: 'ghost',
     class: tw([
       'text-secondary',
-      'enabled:hover:bg-secondary-container-hovered',
-      'enabled:active:bg-secondary-container-hovered',
-      'focus-visible:bg-secondary-container-focused',
+      'enabled:hover:bg-secondary/dim-5',
+      'enabled:active:bg-secondary/dim-5',
+      'focus-visible:bg-secondary/dim-5',
     ]),
   },
   {
@@ -26,9 +26,9 @@ export const ghostVariants = [
     design: 'ghost',
     class: tw([
       'text-success',
-      'enabled:hover:bg-success-container-hovered',
-      'enabled:active:bg-success-container-hovered',
-      'focus-visible:bg-success-container-focused',
+      'enabled:hover:bg-success/dim-5',
+      'enabled:active:bg-success/dim-5',
+      'focus-visible:bg-success/dim-5',
     ]),
   },
   {
@@ -36,9 +36,9 @@ export const ghostVariants = [
     design: 'ghost',
     class: tw([
       'text-alert',
-      'enabled:hover:bg-alert-container-hovered',
-      'enabled:active:bg-alert-container-hovered',
-      'focus-visible:bg-alert-container-focused',
+      'enabled:hover:bg-alert/dim-5',
+      'enabled:active:bg-alert/dim-5',
+      'focus-visible:bg-alert/dim-5',
     ]),
   },
   {
@@ -46,9 +46,9 @@ export const ghostVariants = [
     design: 'ghost',
     class: tw([
       'text-error',
-      'enabled:hover:bg-error-container-hovered',
-      'enabled:active:bg-error-container-hovered',
-      'focus-visible:bg-error-container-focused',
+      'enabled:hover:bg-error/dim-5',
+      'enabled:active:bg-error/dim-5',
+      'focus-visible:bg-error/dim-5',
     ]),
   },
   {
@@ -56,19 +56,19 @@ export const ghostVariants = [
     design: 'ghost',
     class: tw([
       'text-neutral',
-      'enabled:hover:bg-neutral-container-hovered',
-      'enabled:active:bg-neutral-container-hovered',
-      'focus-visible:bg-neutral-container-focused',
+      'enabled:hover:bg-neutral/dim-5',
+      'enabled:active:bg-neutral/dim-5',
+      'focus-visible:bg-neutral/dim-5',
     ]),
   },
   {
     intent: 'surface',
     design: 'ghost',
     class: tw([
-      'text-on-surface',
-      'enabled:hover:bg-surface-hovered',
-      'enabled:active:bg-surface-hovered',
-      'focus-visible:bg-surface-focused',
+      'text-surface',
+      'enabled:hover:bg-surface/dim-5',
+      'enabled:active:bg-surface/dim-5',
+      'focus-visible:bg-surface/dim-5',
     ]),
   },
 ] as const