+++ packages/components/checkbox/src/Checkbox.tsx
@@ -1,52 +1,17 @@
 import * as CheckboxPrimitive from '@radix-ui/react-checkbox'
-import { Check } from '@spark-ui/icons/src/icons/Check'
-import React, { ReactNode } from 'react'
+import React from 'react'
 
-import { styles, type StylesProps } from './Checkbox.variants'
+import { Input, type InputProps } from './CheckboxInput'
+import { Label } from './CheckboxLabel'
 
-interface CheckboxRadixProps {
-  /**
-   * The checked icon to use
-   */
-  icon?: ReactNode
-  /**
-   * The checked state of the checkbox when it is initially rendered. Use when you do not need to control its checked state.
-   */
-  defaultChecked?: boolean
-  /**
-   * The controlled checked state of the checkbox. Must be used in conjunction with onCheckedChange.
-   */
-  checked?: boolean
-  /**
-   * Event handler called when the checked state of the checkbox changes.
-   */
-  onCheckedChange?: (checked: boolean) => void
-  /**
-   * When true, prevents the user from interacting with the checkbox.
-   */
-  disabled?: boolean
-  /**
-   * When true, indicates that the user must check the checkbox before the owning form can be submitted.
-   */
-  required?: boolean
-  /**
-   * The name of the checkbox. Submitted with its owning form as part of a name/value pair.
-   */
-  name?: string
-}
-
-export interface CheckboxProps
-  extends CheckboxRadixProps, // Radix props
-    StylesProps, // CVA props (variants)
-    Omit<React.ButtonHTMLAttributes<HTMLButtonElement>, 'value'> {} // Native HTML props
+export type CheckboxProps = InputProps
 
 export const Checkbox = React.forwardRef<
   React.ElementRef<typeof CheckboxPrimitive.Root>,
   CheckboxProps
->(({ intent, icon = <Check />, ...props }, ref) => (
-  <CheckboxPrimitive.Root ref={ref} className={styles({ intent })} {...props}>
-    <CheckboxPrimitive.Indicator className="text-surface flex items-center justify-center">
-      {icon}
-    </CheckboxPrimitive.Indicator>
-  </CheckboxPrimitive.Root>
+>(({ children, ...props }, ref) => (
+  <Label disabled={props.disabled}>
+    <Input ref={ref} {...props} />
+    {children}
+  </Label>
 ))