+++ packages/components/checkbox/src/CheckboxInput.styles.ts
@@ -1,6 +1,7 @@
+import { makeVariants } from '@spark-ui/internal-utils'
 import { cva, VariantProps } from 'class-variance-authority'
 
-export const styles = cva(
+export const inputStyles = cva(
   [
     'h-sz-20 w-sz-20 border-md peer items-center justify-center rounded-sm bg-transparent',
     'spark-disabled:opacity-dim-3 spark-disabled:cursor-not-allowed spark-disabled:hover:ring-0',
@@ -9,7 +10,7 @@ export const styles = cva(
   ],
   {
     variants: {
-      intent: {
+      intent: makeVariants<'intent', ['primary', 'success', 'alert', 'error']>({
         primary: [
           'spark-state-unchecked:border-outline',
           'spark-state-checked:border-primary spark-state-checked:bg-primary',
@@ -26,7 +27,7 @@ export const styles = cva(
           'spark-state-unchecked:border-error',
           'spark-state-checked:border-error spark-state-checked:bg-error',
         ],
-      },
+      }),
     },
     defaultVariants: {
       intent: 'primary',
@@ -34,4 +35,4 @@ export const styles = cva(
   }
 )
 
-export type StylesProps = VariantProps<typeof styles>
+export type InputStylesProps = VariantProps<typeof inputStyles>