+++ packages/components/visually-hidden/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/adevinta/spark/compare/@spark-ui/visually-hidden@0.3.0...@spark-ui/visually-hidden@0.4.0) (2023-04-12)
+
+### Features
+
+- **visually-hidden:** froze versions ([a209985](https://github.com/adevinta/spark/commit/a209985c0a3506e34c7f57888d8fda09452be0b2))
+
+# [0.3.0](https://github.com/adevinta/spark/compare/@spark-ui/visually-hidden@0.2.0...@spark-ui/visually-hidden@0.3.0) (2023-04-07)
+
+### Features
+
+- add peerDeps field ([18e758d](https://github.com/adevinta/spark/commit/18e758d4796389711040fed1c9b738270c505abf))
+
+# [0.2.0](https://github.com/adevinta/spark/compare/@spark-ui/visually-hidden@0.1.3...@spark-ui/visually-hidden@0.2.0) (2023-04-07)
+
+### Features
+
+- update all dependencies to use caret versions in every packages ([35e329c](https://github.com/adevinta/spark/commit/35e329c39bdc661f477d22e770d82e72d7f93a75))
+
 ## [0.1.3](https://github.com/adevinta/spark/compare/@spark-ui/visually-hidden@0.1.2...@spark-ui/visually-hidden@0.1.3) (2023-04-06)
 
 **Note:** Version bump only for package @spark-ui/visually-hidden